feat: logic fully extracted
