+++ cmd/logic/cron_set_test.go
@@ -1,4 +1,4 @@
-package cmd
+package logic
 
 import (
 	"errors"
