+++ cmd/logic/version_upgrade_exec.go
@@ -1,0 +1,8 @@
+//go:build !windows
+// +build !windows
+
+package logic
+
+func upgradeWin(versionFlag string) error {
+	return nil
+}
