+++ cmd/login.go
@@ -1,9 +1,7 @@
 package cmd
 
 import (
-	"fmt"
-
-	"github.com/deta/deta-cli/runtime"
+	"github.com/deta/deta-cli/cmd/logic"
 	"github.com/spf13/cobra"
 )
 
@@ -20,26 +18,5 @@ func init() {
 }
 
 func login(cmd *cobra.Command, args []string) error {
-	fmt.Println("Please, log in from the web page. Waiting...")
-	if err := authManager.Login(); err != nil {
-		return err
-	}
-
-	u, err := client.GetUserInfo()
-	if err != nil {
-		return err
-	}
-
-	runtimeManager, err := runtime.NewManager(nil, false)
-	if err != nil {
-		return err
-	}
-
-	runtimeManager.StoreUserInfo(&runtime.UserInfo{
-		DefaultSpace:     u.DefaultSpace,
-		DefaultSpaceName: u.DefaultSpaceName,
-		DefaultProject:   u.DefaultProject,
-	})
-	fmt.Println("Logged in successfully.")
-	return nil
+	return logic.Login(client, authManager, args)
 }
