+++ cmd/version_upgrade.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"fmt"
-	"os/exec"
-	"runtime"
-	"strings"
-
+	"github.com/deta/deta-cli/cmd/logic"
 	"github.com/spf13/cobra"
 )
 
@@ -27,68 +23,7 @@ func init() {
 }
 
 func upgrade(cmd *cobra.Command, args []string) error {
-	latest, err := getLatestVersion()
-	if err != nil {
-		return err
-	}
-
-	upgradingTo := latest.Tag
-	if versionFlag != "" {
-		if !strings.HasPrefix(versionFlag, "v") {
-			versionFlag = fmt.Sprintf("v%s", versionFlag)
-		}
-
-		versionExists, err := checkVersionExists(versionFlag)
-		if err != nil {
-			return err
-		}
-		if !versionExists {
-			return fmt.Errorf("no such version")
-		}
-
-		upgradingTo = versionFlag
-	}
-	if detaVersion == upgradingTo {
-		fmt.Printf("Version already %s, no upgrade required\n", upgradingTo)
-		return nil
-	}
-
-	switch runtime.GOOS {
-	case "linux", "darwin":
-		return upgradeUnix()
-	case "windows":
-		return upgradeWin()
-	default:
-		return fmt.Errorf("unsupported platform")
-	}
-}
-
-func upgradeUnix() error {
-	curlCmd := exec.Command("curl", "-fsSL", "https://get.deta.dev/cli.sh")
-	msg := "Upgrading deta cli"
-	curlOutput, err := curlCmd.CombinedOutput()
-	if err != nil {
-		fmt.Println(string(curlOutput))
-		return err
-	}
-
-	co := string(curlOutput)
-	shCmd := exec.Command("sh", "-c", co)
-	if versionFlag != "" {
-		if !strings.HasPrefix(versionFlag, "v") {
-			versionFlag = fmt.Sprintf("v%s", versionFlag)
-		}
-		msg = fmt.Sprintf("%s to version %s", msg, versionFlag)
-		shCmd = exec.Command("sh", "-c", co, "upgrade", versionFlag)
-	}
-	fmt.Printf("%s...\n", msg)
-
-	shOutput, err := shCmd.CombinedOutput()
-	fmt.Println(string(shOutput))
-	if err != nil {
-		return err
-	}
-	return nil
+	return logic.Upgrade(client, versionFlag, args)
 }
 
 func versionUpgradeExamples() string {
