--- cmd/version_upgrade_exec.go
@@ -1,7 +1,0 @@
-// +build !windows
-
-package cmd
-
-func upgradeWin() error {
-	return nil
-}
